ai_services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langchain==0.3.13
langchain-mistralai==0.2.4
pydantic==2.10.5
llama-index>=0.10.0
llama-index-llms-mistralai
llama-index-embeddings-mistralai
scipy
numpy
